openapi-request-coercer: fix header coercing (#802)
Header field names are case insensitive which was addressed to some
extent in the code by lowercasing the properties of the `coercers`
object. However, the function performing the coercion compared the
original (non-lowercase) fields to the lowercase coercers properties,
which doesn't work unless the header was lowercase to begin with.

This is fixed here by using the same lowercasing on the parameter name
when checking for a matching coercer.