Fix dynamic import to work with jest
When using jest this code gets transpiled to `require` and the file URL syntax of the import is not supported and breaks the testing environment.  This change only uses file URL import syntax on windows machines.  This fixes our tests and is still supported on linux and mac operating systems.

I think windows is definitely in the minority here and we should not optimize for it.  Furthermore, I believe using a UNC path would be a better solution outright but I cannot test because I don't have a window machine at hand.

Why not just use jest's `--experimental-vm-modules` flag? Glad you asked.  Because of this [jest bug](https://github.com/jestjs/jest/issues/11434#issuecomment-1410445848) which is caused by this [node bug](https://github.com/nodejs/node/issues/37648) which is ultimately caused by this [v8 bug](https://issues.chromium.org/issues/40784051)